Show empty fence when a building doesn't exist in the present
This feature of the time evolution will add empty buildings children
to a quarter when its data (the file in the examples) doesn't exist
in the present but used to exist in the past. This allow the
application to make a time evolution backwards, it means that
the empty zones showed in the present will be filled in the past
time snapshots.

The dataframes has now more enriched information in order to have
this feature, the backups have been updated.

The examples of Perceval and Angular have been updated.

Signed-off-by: David Moreno <dmoreno@bitergia.com>