Add --commitbycommit arg to do commit time evolution
This new argument will be used instead of samples argument. It will
return a main json with the information that used to have and a
new flag saying that there is a commit by commit evolution.

Moreover, there will be a commit list and there will be the commit
sha information with the date in the data_list array.

Signed-off-by: David Moreno <dmoreno@bitergia.com>