Save all the time evolution data in the main_json file
In order to avoid multiple JSON petitions from the client, this
commit changes the strucuture of the files adding all the
information in the main_json file. There is an argument
called `--exported-snapshots`, if activated the files
will be saved as before.

time-deltas argument added in order to difference between the
two time evolution analisys that we have.

Signed-off-by: David Moreno <dmoreno@bitergia.com>