Fix ssldev so that it works with OpenSSL in recent Debian versions
When running "npm ssldev" in recent Debian/testing versions, it
fails because OpenSSL is configured so that it needs larger
keys than those provided by the public-encrypt package (in its
testing directory). The fix is to tell OpenSSL to ignore that
default which should be enough for testing purposes. The true
fix would be using larger keys, but that is not in this patch.