[bars] Fix how Y axis is being constructed
The code for building the Y axis was a bit messy, with a loop
creaating a lot of labels. Instead of that, run a loop only for
the places where a tick should be shown.

In addition, this fixes a problem with large values for the Y axis:
with 50,000, for example, the page just got wild (maybe trying to
create too many elements in the scene). Now, it seems to work.