Add legend on top of node upon intersection with mouse raycaster
This commit changes the raycaster from the center of the scene
to the mouse. It deletes the functionality that added legends
on top of the scene and adds a new one that shows a legend on top
of a node when the cursor intersects with it. It also adds a
functionality that removes this legend when the intersection
changes and the cursor leaves the node. These legends turn to
follow the position of the user, so that their content is always
visible. They also change their height according to the radius
of the node sphere.