[bars] New example and fix for it work
The new example I added (bars_change.html) for bars,
didn't work. The example draw some bars charts,
and then updates them twice, with new data for some of
the charts. It worked well for the first data, but
when it was updated, apparently it didn't use the new
data. Wne it was updated for the second time, apparently
it used the data from the first change.

Looking at the data, it seems self.currentData was updated
too late, after updating the chart (and thus, the chart
was only updated the next time update is called).

The fix is to first update self.currentData, and only then
update the chart.

I hope this does not break other stuff