[selector] New BabiaXR component, for selecting some data
The selector component is used to select from a list of data
items, those that match, for a certain field name, a certain
value. In fact, selector first groups the items by the
field used for the selection, and then dispatches data
to other components interested in it. Each time it dispatches
data, it corresponds to the items for a certain value
of the field used for the selection.

An example is included.

The implementation is suboptimal, but I expect to refactor it
later.