Add multiuser demo
This commit adds a dependency for networked-aframe. It is used to
create an example of a multiuser scene. This example contains audio,
syncronization and persistance of entities. It also allows exchange of
messages between clients. It contains a login screen where serverURL
for WebRTC server can be defined, as well as a username and an avatar
color.