Add names for axis
This commit adds names for the three axis of babia-axis and adds
an attribute in babia-bars that lets us choose if we want to show
the names (name set to true) or not (name set to false, by default).
It also adds a simple example that shows babia-bars with names on axis.