Add axis names to barsmap, update docs and add examples
This commit adds the attribute axis_name to babia-barsmap component,
allowing us to decide if we want to add the names of the metrics (true)
or not (false). It also updates the docs for babia-bars and babia-barsmap,
adding information about axis_name. It adds, as well, the API documentation
for the babia-axis components (x, y and z). Finally, an example for
babia-barsmap has been added and the babia-bars example has been moved
to the right folder to keep the proper logic since it uses a querier. Two
new screenshot tests have been added to get the photos for our examples
page for these two examples.