Add async-filter and simplify all async code
This commit adds the new async-filter component and simplifies all
async code, this way most of the code is inside the register and
the components are much cleaner and easier to understand. We also
avoid duplicated code.

It also adds an example with async-filter (with async-querier and
async-visualizer) to test its behaviour.