Add promisedData interval, rename functions and parameterize time
This commit fixes the promisedData promise, by adding an interval
that allows the code to keep running while it tries to retrieve the
data. It also adds a maximum number of trials.

Some functions have been renamed in order to clean and refractor
the code in an easier way and time and max (maximum number of trials)
for the promisedData have been parameterized.