Fix bug in babia-network and update examples
This commit fixes a bug that did not allow to find the querier or
filter component when it was not specified inside 'from'. It has been
fixed by deleting the 'undefined' default value for 'from', which
allows to go through all the possible from components in the first
update.

This commit also reorganizes the network examples and adds the missing
screenshot tests.