Add 2-sided communication with notiBuffer and add it to components
This commit adds notiBuffer to the following components:
babia-selector, babia-navigator (with babia-slider, babia-controls,
babia-step-controller and babia-speed-controller). For the two-sided
communication with notiBuffer, parameters such as functions or details
have been added to the constructor and methods in notiBuffer, so
that all components can find each other and react to registrations
and unregistrations. This commit also cleans some code in
babia-filter, babia-bars and babia-barsmap and adds some functions
to common.js to simplify. Finally, some examples have been
adjusted and/or corrected.