Update docs and examples changing producer-consumer order
This commit updates the docs relating NotiBuffer to clarify that
the entity containing the producer must be written before the entity
containing the consumer. In case they are both in the same entity,
there is not need to be in an specific order. This commit also updates
the babia-boats examples with selector so they navComponent is properly
found in the selector.