Add current_value, step, speed and state attributes
This commit adds the necessary attributes to babia-selector,
babia-navigator, babia-controls and babia-sliders (the three of
them) in order to update and react when their attributes are
modified. This behaviour allows a multiuser connection by sharing
only selector and navigator. The multi_boats with_selector example
has been modified so that it works with the current changes. In it,
the "not owner" users only see the main slider, until they touch it,
then they become the owners.