Add multiuser mode inside the babia-selector component
This commit adds the multiuser mode inside the selector component,
this way we only have to do the regular steps to share the selector
in the scene and it will activate the multiuser behaviour
automatically, with no extra code for the ownership and interaction
logic and no need for sharing the navigator.

This commit also updates the examples and the documentation.