boats: New parameter diffBuildingHeight
Height of buildings was calculated relative to the difference between
the maximum and the minimum value of (height) data.
This is a problem when the minimum value is not 0, as is the case
in the  boats2 example, where there are several data with height 1,
and none with height 0.

diffBuildingHeight, when set to true, follows this "old" behavior,
since it could be useful in some cases. But if set to false, height
is calculated relative to the difference between 0 and the maximum data.
By default, it is set to false, since this seems the most usual behavior.

There is also some (small) refactoring in this commit, to make some
code (hopefully) a bit easier to understand.