Ensure this.selectable is correctly initialized
We need this.selectable to be always present, even if no data
was still assigned to it. Otherwise, some parts of the code
may try to access properties of this.selectable, and fail if
it is null.

Ensuring this.selectable is aways present includes separating
the update of its data (list) from the other parameters
(field, speed, step, etc.). The data should be set when it is
available (in processData), while the others can be set in the
init() of the component.